requirements.txt
python-dotenv>=1.0.0
boto3>=1.34.0
numpy>=1.26.0
pydantic>=2.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
cerebras-cloud-sdk>=1.0.0
httpx>=0.25.0
psycopg2-binary>=2.9.9
